save corpus for existing corpus

+++ edu.usc.cssl.tacit.common/src/edu/usc/cssl/tacit/common/corpusmanagement/manageCorpora.java
@@ -2,22 +2,22 @@ package edu.usc.cssl.tacit.common.corpusmanagement;
 
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 import org.apache.commons.io.FileUtils;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-
-import edu.usc.cssl.tacit.common.ui.corpusmanagement.internal.ICorpus;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
 import edu.usc.cssl.tacit.common.ui.corpusmanagement.internal.ICorpusClass;
 import edu.usc.cssl.tacit.common.ui.corpusmanagement.services.Corpus;
 import edu.usc.cssl.tacit.common.ui.corpusmanagement.services.CorpusClass;
 
-public class ManageCorpora {
+public class manageCorpora {
 
 	static String rootDir = System.getProperty("user.dir")
 			+ System.getProperty("file.separator") + "tacit_corpora"
@@ -91,6 +91,22 @@ public class ManageCorpora {
 			String metaFile = corpusLocation
 					+ System.getProperty("file.separator") + "meta.txt";
 
+			JSONParser parser = new JSONParser();
+			JSONArray analysisArray = new JSONArray();
+			int numAnalysis = 0;
+			try {
+				analysisArray = (JSONArray) ((JSONObject) parser
+						.parse(new FileReader(metaFile))).get("prev_analysis");
+				numAnalysis = (Integer) ((JSONObject) parser
+						.parse(new FileReader(metaFile))).get("num_analysis");
+			} catch (FileNotFoundException e) {
+				e.printStackTrace();
+			} catch (IOException e) {
+				e.printStackTrace();
+			} catch (ParseException e) {
+				e.printStackTrace();
+			}
+
 			File metaFp = new File(metaFile);
 
 			JSONObject jsonObj = new JSONObject();
@@ -104,6 +120,36 @@ public class ManageCorpora {
 					.getClasses();
 
 			JSONArray classArray = new JSONArray();
+
+			for (int i = 0; i < numClasses; i++) {
+				CorpusClass currClass = (CorpusClass) corporaClasses.get(i);
+				JSONObject classObj = new JSONObject();
+				classObj.put("class_name", currClass.getClassName());
+				classObj.put("original_loc", currClass.getClassPath());
+
+				String[] dirParts = currClass.getClassName().split(
+						System.getProperty("file.separator"));
+				String dirName = dirParts[dirParts.length - 1];
+				classObj.put("tacit_loc",
+						corpusLocation + System.getProperty("file.separator")
+								+ dirName);
+
+				classArray.add(classObj);
+			}
+
+			jsonObj.put("class_details", classArray);
+			jsonObj.put("num_analysis", numAnalysis);
+			jsonObj.put("prev_analysis", analysisArray);
+
+			try {
+				BufferedWriter bw = new BufferedWriter(new FileWriter(metaFp));
+				bw.write(jsonObj.toString());
+				bw.close();
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+			
+			copyCorpus(jsonObj);
 		}
 
 	}
@@ -118,43 +164,16 @@ public class ManageCorpora {
 			String tacitLoc = (String) ((JSONObject) classArray.get(i))
 					.get("tacit_loc");
 
-			new File(tacitLoc).mkdir();
+			if (!(new File(tacitLoc).exists())) {
+				new File(tacitLoc).mkdir();
 
-			try {
-				FileUtils.copyDirectory(new File(originalLoc), new File(
-						tacitLoc));
-			} catch (IOException e) {
-				e.printStackTrace();
+				try {
+					FileUtils.copyDirectory(new File(originalLoc), new File(
+							tacitLoc));
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
 			}
 		}
 	}
-
-	public static String[] getNames() {
-		List<ICorpus> readCorpusList = readCorpusList();
-		List<String> names = new ArrayList<String>();
-		for (ICorpus iCorpus : readCorpusList) {
-			names.add(iCorpus.getCorpusId());
-		}
-		return (String[]) names.toArray(new String[names.size()]);
-	}
-
-	public static List<ICorpus> readCorpusList() {
-
-		// TO-DO
-		// Complete the implementation
-		return Collections.EMPTY_LIST;
-
-	}
-
-	public static ICorpus readCorpusById(String id) {  // why? - to get updated corpus instead of stale data at tool
-
-		List<ICorpus> readCorpusList = readCorpusList();
-		for (ICorpus iCorpus : readCorpusList) {
-			if (iCorpus.getCorpusId().equals(id)) {
-				return iCorpus;
-			}
-		}
-		return null;
-	}
-
 }
